+++ components/header/Header.jsx
@@ -1,7 +1,7 @@
 import { useState } from 'react';
 import { SafeAreaView, TouchableOpacity, View, Text, Image, Modal } from 'react-native';
 import styles from './header.style';
-import icons from '../../../constants/icons';
+import icons from '../../constants/icons';
 
 // This is a custom header
 // Users can click on the menu button to toggle a modal with a logout button
