+++ components/index.js
@@ -1,9 +1,9 @@
 // export all components from here
-import Login from "./home/login/Login";
-import Welcome from "./home/welcome/Welcome";
-import Chores from "./home/chores/Chores";
-import Header from "./home/header/Header";
-import ReminderPicker from "./home/chores/reminderPicker/ReminderPicker";
+import Login from "./login/Login";
+import Welcome from "./welcome/Welcome";
+import Chores from "./chores/Chores";
+import Header from "./header/Header";
+import ReminderPicker from "./reminderPicker/ReminderPicker";
 
 export {
     Login,
