+++ components/welcome/welcome.style.js
